% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pag.R
\name{plot_pag}
\alias{plot_pag}
\title{Plot a PAG returned by PFCI}
\usage{
plot_pag(fit, ...)
}
\arguments{
\item{fit}{A pfci_fit object returned by pfci_fit(), or a pcalg fci object.}

\item{...}{Additional arguments passed to pcalg plotting.}
}
\description{
Plot a PAG returned by PFCI
}
